ml_model CRUC funcs created
