+++ package.json
@@ -20,6 +20,7 @@
   },
   "homepage": "https://github.com/seahahn/User-Proj-Managing#readme",
   "dependencies": {
+    "aws-sdk": "^2.1091.0",
     "dotenv": "^16.0.0",
     "fastify": "^3.27.2",
     "fastify-cli": "^2.15.0",
@@ -29,6 +30,7 @@
     "fastify-plugin": "^3.0.1",
     "fastify-postgres": "^3.6.0",
     "fastify-swagger": "^4.15.0",
+    "lodash": "^4.17.21",
     "pg": "^8.7.3"
   }
 }
