+++ src/schemas/index.js
@@ -1,0 +1,22 @@
+import { saveProject, loadProject, updateProject, deleteProject, updateProjName, getProjList } from "./mlProject/index.js";
+import {
+    saveModel,
+    getModelList,
+    updateModelName,
+    deleteModel,
+    // loadModel, updateModel, , ,
+} from "./mlModel/index.js";
+
+export {
+    saveProject,
+    loadProject,
+    updateProject,
+    deleteProject,
+    updateProjName,
+    getProjList,
+    saveModel,
+    getModelList,
+    updateModelName,
+    deleteModel,
+    // loadModel, updateModel, , ,
+};
