+++ src/schemas/mlModel/index.js
@@ -1,0 +1,14 @@
+import saveModel from "./saveModel.schema.js";
+// import loadModel from "./loadModel.schema.js";
+// import updateModel from "./updateModel.schema.js";
+import deleteModel from "./deleteModel.schema.js";
+import updateModelName from "./updateModelName.schema.js";
+import getModelList from "./getModelList.schema.js";
+
+export {
+    saveModel,
+    getModelList,
+    updateModelName,
+    deleteModel,
+    // loadModel, updateModel, , ,
+};
